opt/lxd-ddns/requirements.txt
dnspython3>=1.12.0
pylxd>=2.2.0
